### 2025-09-22 #9053 @editorialbot
[PRE REVIEW]: Xbitinfo: Compressing geospatial data based on information theory
**Submitting author:** @observingClouds (<a href="http://orcid.org/0000-0001-5468-1137">Hauke Schulz</a>)
**Repository:** https://github.com/observingClouds/xbitinfo
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.0.4
