### 2025-09-22 #9053 @danielskatz
@observingClouds - thanks for your submission. I'll look for an editor next. While I do so, please notice that a couple of the references have misinterpreted authors, where the author field in the bibtex should be protected by double braces, e.g. `{{full author name}}`
