### 2025-10-10 #9178 @editorialbot
[REVIEW]: Xbitinfo: Compressing geospatial data based on information theory
**Submitting author:** @observingClouds (<a href="http://orcid.org/0000-0001-5468-1137">Hauke Schulz</a>)
**Repository:** https://github.com/observingClouds/xbitinfo
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.0.6
