### 2025-10-31 9a68915
Mention pooch requirement for example dataset
Fixes #312
