### 2025-10-31 0e87082
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
