+++ README.md
@@ -54,16 +54,21 @@ pip install xbitinfo  # julia needs to be installed manually
 ```python
 import xarray as xr
 import xbitinfo as xb
-example_dataset = 'eraint_uvz'
+
+example_dataset = "eraint_uvz"
 ds = xr.tutorial.load_dataset(example_dataset)
 # Step 1: analyze bitwise information content
 bitinfo = xb.get_bitinformation(ds, dim="longitude")
 
 # Step 2: decide on a threshold of real information to preserve (e.g. 99%)
-keepbits = xb.get_keepbits(bitinfo, inflevel=0.99)  # get number of mantissa bits to keep for 99% real information
+keepbits = xb.get_keepbits(
+    bitinfo, inflevel=0.99
+)  # get number of mantissa bits to keep for 99% real information
 
 # Step 3: reduce the precision of the dataset accordingly (bitrounding)
-ds_bitrounded = xb.xr_bitround(ds, keepbits)  # bitrounding keeping only keepbits mantissa bits
+ds_bitrounded = xb.xr_bitround(
+    ds, keepbits
+)  # bitrounding keeping only keepbits mantissa bits
 
 # Step 4: apply lossless compression (e.g. zlib, blosc, zstd) and store the dataset
 ds_bitrounded.to_compressed_netcdf(outpath)
