+++ CHANGELOG.rst
@@ -2,6 +2,11 @@
 CHANGELOG
 =========
 
+unreleased
+------------------
+
+* Update pre-commit hooks and update doctests (:pr:`307`) `Hauke Schulz`_.
+
 0.0.4 (2025-09-22)
 ------------------
 
