+++ xbitinfo/xbitinfo.py
@@ -621,24 +621,24 @@ def get_keepbits(info_per_bit, inflevel=0.99, information_filter=None, **kwargs)
     <xarray.Dataset> Size: 28B
     Dimensions:   (inflevel: 1)
     Coordinates:
-        dim       <U3 12B 'lon'
       * inflevel  (inflevel) float64 8B 0.99
+        dim       <U3 12B 'lon'
     Data variables:
         air       (inflevel) int64 8B 7
     >>> xb.get_keepbits(info_per_bit, inflevel=0.99999999)
     <xarray.Dataset> Size: 28B
     Dimensions:   (inflevel: 1)
     Coordinates:
-        dim       <U3 12B 'lon'
       * inflevel  (inflevel) float64 8B 1.0
+        dim       <U3 12B 'lon'
     Data variables:
         air       (inflevel) int64 8B 7
     >>> xb.get_keepbits(info_per_bit, inflevel=1.0)
     <xarray.Dataset> Size: 28B
     Dimensions:   (inflevel: 1)
     Coordinates:
-        dim       <U3 12B 'lon'
       * inflevel  (inflevel) float64 8B 1.0
+        dim       <U3 12B 'lon'
     Data variables:
         air       (inflevel) int64 8B 52
     >>> info_per_bit = xb.get_bitinformation(ds)
