### 2025-10-31 346d650
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
