### 2025-10-31 1e86bb1
Mention pooch requirement for example dataset (#313)
* Mention pooch requirement for example dataset

Fixes #312

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
