### 2025-11-07 f248e91
Update CHANGELOG with recent changes
