+++ .github/workflows/ci.yaml
@@ -24,10 +24,19 @@ jobs:
         with:
           keyword: '[skip-ci]'
   test:
-    name: pytest
+    name: pytest on Python ${{ matrix.python-version }}
     runs-on: ubuntu-latest
     needs: detect-ci-trigger
     if: needs.detect-ci-trigger.outputs.triggered == 'false'
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version:
+          - '3.14'
+          - '3.13'
+          - '3.12'
+          - '3.11'
+          - '3.10'
     defaults:
       run:
         shell: 'bash -l {0}'
@@ -42,6 +51,7 @@ jobs:
         uses: astral-sh/setup-uv@v5.4.1
         with:
           enable-cache: true
+          python-version: ${{ matrix.python-version }}
           cache-dependency-glob: "**/pyproject.toml"
       - name: Run tests
         run: uv run --group test pytest tests
