+++ pyproject.toml
@@ -3,7 +3,7 @@ name = "xbitinfo"
 dynamic = ["version"]
 description = "Retrieve information content and compress accordingly."
 readme = "README.md"
-requires-python = ">=3.9"
+requires-python = ">=3.9,<3.14"
 license = { text = "MIT license" }
 keywords = ["xbitinfo"]
 dependencies = [
