### 2025-11-07 32abd23
allow higher python versions in test
