### 2025-11-07 70606e0
add compatibility to zarr v2
