+++ .github/workflows/ci.yaml
@@ -50,9 +50,13 @@ jobs:
       - name: Install uv
         uses: astral-sh/setup-uv@v5.4.1
         with:
-          enable-cache: true
+          enable-cache: false
           python-version: ${{ matrix.python-version }}
           cache-dependency-glob: "**/pyproject.toml"
+      - name: Install dependencies
+        run: uv sync --group test
+      - name: List installed packages
+        run: uv run --group test pip list
       - name: Run tests
         run: uv run --group test pytest tests
   doctest:
