### 2025-11-07 bce2da1
add zarr<3 support to tests
