### 2025-11-07 0ccaae7
Revert "add zarr<3 support to tests"
This reverts commit bce2da18bd972abcceadbabd2b04cbe326546129.
