### 2025-11-07 4d6e7a3
Revert "add compatibility to zarr v2"
This reverts commit 70606e062dc586e8a77a9ce07958a85b0d3ce1e3.
