### 2025-11-07 7e7c37a
drop python 3.10 support
