+++ .github/workflows/ci.yaml
@@ -32,11 +32,9 @@ jobs:
       fail-fast: false
       matrix:
         python-version:
-          - '3.14'
           - '3.13'
           - '3.12'
           - '3.11'
-          - '3.10'
     defaults:
       run:
         shell: 'bash -l {0}'
