+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ unreleased
 ------------------
 
 * Update pre-commit hooks and update doctests (:pr:`307`) `Hauke Schulz`_.
+* Add python matrix test and restrict supported versions to 3.11-3.13 (:pr:`323`) `Hauke Schulz`_.
 
 0.0.4 (2025-09-22)
 ------------------
