### 2025-11-07 42e1303
CI: python version matrix test + restrict to 3.11-3.13 (#323)
