### 2025-11-07 df511d5
Update CHANGELOG for version 0.0.5
