### 2025-11-07 1f75eae
set outpath to example file
fixes #315
