### 2025-11-07 a08f24b
set implementation to python for simplicity
fixes #314
