### 2025-11-07 #9178 @tomvothecoder
### Issues running latest xbitinfo-python through conda-forge -- out of date?

I am now working on  "Functionality: Have the functional claims of the software been confirmed?".

Specifically, I am running the minimum example for the ["How to use"](https://xbitinfo.readthedocs.io/en/latest/#how-to-use) guide.

1. Successfully created the envronment
	```bash
	conda create -n xbitinfo-test -c conda-forge xbitinfo-python
	conda activate xbitinfo-test
	```
2. Tried to run the script and get:
   ```python
	   Traceback (most recent call last):
	  File "<stdin>", line 1, in <module>
	    import xbitinfo as xb
	  File "/global/u2/v/vo13/tomvothecoder/xbitinfo/xbitinfo/__init__.py", line 10, in <module>
	    from .bitround import jl_bitround, xr_bitround
	  File "/global/u2/v/vo13/tomvothecoder/xbitinfo/xbitinfo/bitround.py", line 4, in <module>
	    from .xbitinfo import _jl_bitround, get_keepbits
	  File "/global/u2/v/vo13/tomvothecoder/xbitinfo/xbitinfo/xbitinfo.py", line 9, in <module>
	    from prefect import flow, task, unmapped
	ImportError: cannot import name 'flow' from 'prefect' (/global/homes/v/vo13/miniforge3/envs/xbitinfo/lib/python3.14/site-packages/prefect/__init__.py). Did you mean: 'Flow'?
   ```

I noticed the latest version of xbitinfo-python is `0.0.4dev1` on [conda-forge](https://anaconda.org/conda-forge/xbitinfo-python) was released 11 months ago, but on GitHub [v0.04](https://github.com/observingClouds/xbitinfo/releases/tag/v0.0.4) was released in September and [v0.0.5](https://github.com/observingClouds/xbitinfo/releases/tag/v0.0.5) released a few hours ago.

**I think the version on conda-forge is out of date?** The `prefect` dependency may also need to be constrained appropriately to ensure the `ImportError` doesn't appear.
