### 2025-11-08 8e8e7e0
typo and grammar updates
