+++ paper.md
@@ -32,25 +32,23 @@ Xbitinfo analyses datasets based on their bitwise real information content and a
 accordingly. Xbitinfo provides additional tools to visualize the information histograms and to make
 informed decisions on the real information threshold that is subsequently used as the preserved precision
 during the compression of arrays of floating-point numbers. In contrast, the false information is rounded
-to zero using bitrounding. Lossless compression subsequently exploits the high compressibility from tailing
-zero mantissa bits. Xbitinfo's functionality supports xarray datasets to interact with a range of common
-input and output dataformats, including all numcodecs compression algorithms.
+to zero using bit rounding. Lossless compression subsequently exploits the high compressibility from trailing
+zero mantissa bits. Xbitinfo's functionality supports xarray datasets and therefore a range of common
+input and output data formats, and all numcodecs compression algorithms.
 
 # Statement of need
 
 The geosciences, similar to other research fields, are generating more and more data both through
-simulations and observations. At the same time, data storage solutions have not increased at the
-same pace. In addition, more and more data is stored in the cloud and egress fees and network
-speeds are more and more of a concern. Compression algorithms can help to reduce the pressure
+simulations and observations. At the same time, data storage capacity has not kept pace. In addition, data is increasingly stored in the cloud where egress fees and network
+speeds are a concern. Compression algorithms can help to reduce the pressure
 on these components significantly and are therefore commonly used.
 
 Lossless compressors like Zlib or Zstandard encode datasets exploiting redundancies without losing
 any information. This is often unnecessarily conservative as not all the bits are meaningful,
-i.e. they do not contain real information. They often encode unnecessarily high precision of
-floating-point numbers, several orders of magnitude higher than the uncertainty of the data
-(arising from e.g. model, numerical, observational or rounding errors) itself. Lossy compression
+i.e. they do not contain real information. They often encode floating-point numbers with unnecessarily high precision, several orders of magnitude higher than the uncertainty of the data
+(arising from, for example, model, numerical, observational or rounding errors) itself. Lossy compression
 is therefore often used, sacrificing bits with little to no real information, and from image and
-audio compression, JPEG and MP3 are two promintent examples. Geospatial data lacks a similarly
+audio compression, JPEG and MP3 are two prominent examples. Geospatial data lacks a similarly
 widely accepted compression standard.
 
 JPEG and MP3 use perceptual models of the human visual and auditory system to decide on whether
@@ -60,49 +58,48 @@ data that still undergoes mathematical operations, like gradients. Commonly used
 data is linear quantization as it is a standard algorithm supported by the GRIB format. It encodes
 the min-max range of the data into evenly, or linearly, spaced quanta and enumerates those using
 integers. The issue with linear quantization is however that it often is not a good mapping for
-geophysical quantities with a more logarithmical distribution. In practice, the number of preserved
+geophysical quantities with a more logarithmic distribution. In practice, the number of preserved
 mantissa bits in the quantization process is often applied to an entire set of variables and dimensions.
-As a consequence, some variables have too little information preserved while others kept too much
+As a consequence, some variables have too little information preserved while others keep too much
 (false) information.
 
-@klower_compressing_2021 has developed an algorithm that can distinguish between real and false
-information content based on information theory. It further allows to set a threshold for the real
-information content that shall be preserved in case additional compression is needed beyond the
+@klower_compressing_2021 have developed an algorithm that can distinguish between real and false
+information content based on information theory. It further allows one to set a threshold for the real
[…]
