### 2025-11-13 daffe2f
Update CHANGELOG with new pytest section
