### 2025-11-13 b39aed6
Remove setup.py from pytest ignore options
repo does not contain a setup.py any longer
