### 2025-11-13 fec33f0
Update dataset loading instructions in notebook (#322)
* Update dataset loading instructions in notebook

Added installation note for s3fs to access the CONUS dataset.

* Improve wording
