### 2025-11-14 888ff5a
Improve demo code (#325)
* set outpath to example file

fixes #315

* set implementation to python for simplicity

fixes #314

* document changes
