+++ README.md
@@ -55,12 +55,15 @@ pip install xbitinfo  # julia needs to be installed manually
 import xarray as xr
 import xbitinfo as xb
 
+# Define output path for compressed dataset
+outpath = "example_bitrounded_compressed.nc"
+
 # Load example dataset
 # (requires pooch to be installed via e.g. `pip install pooch`)
 example_dataset = "eraint_uvz"
 ds = xr.tutorial.load_dataset(example_dataset)
 # Step 1: analyze bitwise information content
-bitinfo = xb.get_bitinformation(ds, dim="longitude")
+bitinfo = xb.get_bitinformation(ds, dim="longitude", implementation="python")
 
 # Step 2: decide on a threshold of real information to preserve (e.g. 99%)
 keepbits = xb.get_keepbits(
