### 2025-11-20 a61f2ba
support zarr v2 and python 3.10
