+++ .github/workflows/ci.yaml
@@ -35,6 +35,7 @@ jobs:
           - '3.13'
           - '3.12'
           - '3.11'
+          - '3.10'
     defaults:
       run:
         shell: 'bash -l {0}'
