### 2025-11-20 4c758f9
fix crc32c dep issue
