+++ pyproject.toml
@@ -10,7 +10,7 @@ dependencies = [
     "dask",
     "xarray",
     "tqdm",
-    "numcodecs>=0.10.0",
+    "numcodecs>=0.10.0,<0.16.4",
     "matplotlib>=3.9.0",
     "prefect>=3.3.0",
     "zarr",
