### 2025-11-20 f43a447
allow for additional printed xr.Dataset info
