### 2025-11-20 7ef9d3e
support zarr v2 and python 3.10 (#327)
* support zarr v2 and python 3.10

* fix crc32c dep issue

* allow for additional printed xr.Dataset info
