+++ pyproject.toml
@@ -3,14 +3,14 @@ name = "xbitinfo"
 dynamic = ["version"]
 description = "Retrieve information content and compress accordingly."
 readme = "README.md"
-requires-python = ">=3.11,<3.14"
+requires-python = ">=3.10,<3.14"
 license = { text = "MIT license" }
 keywords = ["xbitinfo"]
 dependencies = [
     "dask",
     "xarray",
     "tqdm",
-    "numcodecs>=0.10.0",
+    "numcodecs>=0.10.0,<0.16.4",
     "matplotlib>=3.9.0",
     "prefect>=3.3.0",
     "zarr",
