### 2025-11-20 77022fb
Update CHANGELOG with recent changes
