+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ CHANGELOG
 unreleased
 ------------------
 
+* Update GitHub actions (:pr:`306`) `Hauke Schulz`_.
 * Add note to docs on additional requirements running ArtificialInformation filter example (:pr:`322`) `Hauke Schulz`_.
 * Improve README code example (:pr:`325`) `Hauke Schulz`_.
 * Join pytest toml sections to support pytest>=9.0.0 (:pr:`326`) `Hauke Schulz`_.
