+++ .github/workflows/ci.yaml
@@ -47,7 +47,7 @@ jobs:
         with:
           version: '1.10'
       - name: Install uv
-        uses: astral-sh/setup-uv@v5.4.1
+        uses: astral-sh/setup-uv@v6.3.1
         with:
           enable-cache: true
           python-version: ${{ matrix.python-version }}
@@ -68,7 +68,7 @@ jobs:
         with:
           version: '1.10'
       - name: Install uv
-        uses: astral-sh/setup-uv@v5.4.1
+        uses: astral-sh/setup-uv@v6.3.1
         with:
           enable-cache: true
           cache-dependency-glob: "**/pyproject.toml"
@@ -89,7 +89,7 @@ jobs:
         with:
           version: '1.10'
       - name: Install uv
-        uses: astral-sh/setup-uv@v5.4.1
+        uses: astral-sh/setup-uv@v6.3.1
         with:
           enable-cache: true
           cache-dependency-glob: "**/pyproject.toml"
@@ -116,7 +116,7 @@ jobs:
         with:
           fetch-depth: 0
       - name: Setup python
-        uses: actions/setup-python@v5.5.0
+        uses: actions/setup-python@v5.6.0
         with:
           python-version: '3.11'
       - name: Set up Julia
@@ -124,7 +124,7 @@ jobs:
         with:
           version: '1.10'
       - name: Install uv
-        uses: astral-sh/setup-uv@v5.4.1
+        uses: astral-sh/setup-uv@v6.3.1
         with:
           enable-cache: true
           cache-dependency-glob: "**/pyproject.toml"
