### 2025-11-20 dcc4fef
release v0.0.6
