### 2025-11-21 4adf43a
add additional references
