+++ paper.bib
@@ -45,3 +45,60 @@
   year = {1994},
   organization = {ISO}
 }
+
+@software{zarr,
+  author       = {Alistair Miles and
+                  jakirkham and
+                  David Stansby and
+                  Dimitri Papadopoulos Orfanos and
+                  Joe Hamman and
+                  Davis Bennett and
+                  M Bussonnier and
+                  Josh Moore and
+                  Tom Augspurger and
+                  Deepak Cherian and
+                  Max Jones and
+                  Norman Rzepka and
+                  Hannes Spitz and
+                  Sanket Verma and
+                  James Bourbeau and
+                  Andrew Fulton and
+                  Ryan Abernathey and
+                  Gregory Lee and
+                  Mads R. B. Kristensen and
+                  Zain Patel and
+                  Saransh Chopra and
+                  Matthew Rocklin and
+                  Nathan Zimmerman and
+                  Ilan Gold and
+                  AWA DESTINY AGHANGU and
+                  Ian Hunt-Isaak and
+                  Elliott Sales de Andrade and
+                  Martin Durant},
+  title        = {zarr-developers/zarr-python: v3.1.4},
+  month        = nov,
+  year         = 2025,
+  publisher    = {Zenodo},
+  version      = {v3.1.4},
+  doi          = {10.5281/zenodo.17666358},
+  url          = {https://doi.org/10.5281/zenodo.17666358},
+}
+
+@software{netcdf,
+author = {{NSF Unidata} and Davis, Glenn and Rew, Russ and Heimbigner, Dennis and Hartnett, Edward and Fisher, Ward and Others, Many},
+doi = {10.5065/D6H70CW6},
+license = {BSD-3-Clause},
+title = {{NetCDF-C }},
+url = {https://github.com/Unidata/netcdf-c}
+}
+
+@article{xarray,
+author = {Hoyer, Stephan and Joseph, Hamman},
+doi = {10.5334/jors.148},
+journal = {Journal of Open Research Software},
+month = apr,
+number = {1},
+title = {{xarray: N-D labeled Arrays and Datasets in Python}},
+volume = {5},
+year = {2017}
+}
