### 2025-11-21 7e7f2f4
protect author names
