### 2025-11-21 80a5c82
fix references
