+++ paper.bib
@@ -1,6 +1,6 @@
 @article{klower_compressing_2021,
   title = {Compressing Atmospheric Data into Its Real Information Content},
-  author = {{Klöwer, Milan} and Razinger, Miha and Dominguez, Juan J. and {Düben, Peter D.} and Palmer, Tim N.},
+  author = {{Klöwer}, Milan and Razinger, Miha and Dominguez, Juan J. and {Düben}, Peter D. and Palmer, Tim N.},
   date = {2021-11},
   journaltitle = {Nature Computational Science},
   shortjournal = {Nat Comput Sci},
@@ -17,7 +17,7 @@
 @unpublished{schulz_xbitinfo_2022-1,
   title = {Xbitinfo: {{Compress}} Datasets Based on Their Information Content},
   shorttitle = {Xbitinfo},
-  author = {Schulz, Hauke and Spring, Aaron and {Klöwer, Milan}},
+  author = {Schulz, Hauke and Spring, Aaron and {Klöwer}, Milan},
   date = {2022-10-27},
   doi = {10.5281/zenodo.7259092},
 }
@@ -90,6 +90,7 @@ doi = {10.5065/D6H70CW6},
 license = {BSD-3-Clause},
 title = {{NetCDF-C }},
 url = {https://github.com/Unidata/netcdf-c}
+year = {2025}
 }
 
 @article{xarray,
