+++ paper.bib
@@ -89,7 +89,7 @@ author = {{NSF Unidata} and Davis, Glenn and Rew, Russ and Heimbigner, Dennis an
 doi = {10.5065/D6H70CW6},
 license = {BSD-3-Clause},
 title = {{NetCDF-C }},
-url = {https://github.com/Unidata/netcdf-c}
+url = {https://github.com/Unidata/netcdf-c},
 year = {2025}
 }
 
