### 2025-11-21 #9178 @observingClouds
@macrocosme , @tomvothecoder thank you very much for your thorough review and time commitment. It has helped tremendously in ironing-out the software.

I have addressed now all issues and have released version 0.0.6 with the implemented changes on conda-forge and pypi. I have also updated the paper and fixed some references.
