### 2025-11-21 #9178 @observingClouds
@mstimberg I like to include my current affiliation that is unrelated to this work to the paper, like (`now at institution xyz`), but while this works for my co-authors, [it does not work for me](https://github.com/observingClouds/xbitinfo/blob/0ff49c982786a969a5bb7ed069b09294da00d822/paper.md?plain=1#L6), presumably because the notation clashes with the `corresponding` author flag. Any ideas or suggestions on how to implement this?
