### 2025-11-23 f206950
Modify author details and acknowledgments in paper.md
Updated author affiliations and added funding acknowledgment.
