### 2025-11-24 #9178 @mstimberg
Hi again @observingClouds. It turns out that this is a known bug: openjournals/inara#105. For now, there is no good solution, the recommend approach is to use footnotes for everything, including the corresponding author information. I.e., please set `corresponding: false` (or remove it), and add a `^[Corresponding author]` footnote.
