+++ paper.md
@@ -3,13 +3,14 @@ title: 'xbitinfo: Compressing geospatial data based on information theory'
 tags:
   - Python
 authors:
-  - name: Hauke Schulz^[Corresponding author, Now at Danish Meteorological Institute, Copenhagen, Denmark]
+  - name: Hauke Schulz^[Now at Danish Meteorological Institute, Copenhagen, Denmark]
     orcid: 0000-0001-5468-1137
+    corresponding: false
     affiliation: "1, 2"
   - name: Milan Klöwer
     orcid: 0000-0002-3920-4356
     affiliation: 3
-  - name: Aaron Spring^[Now at XING, Hamburg, Germany]
+  - name: Aaron Spring
     orcid: 0000-0003-0216-2241
     affiliation: 4
 affiliations:
