+++ paper.md
@@ -3,10 +3,9 @@ title: 'xbitinfo: Compressing geospatial data based on information theory'
 tags:
   - Python
 authors:
-  - name: Hauke Schulz
+  - name: Hauke Schulz^[Corresponding author, Now at Danish Meteorological Institute, Copenhagen, Denmark]
     orcid: 0000-0001-5468-1137
-    corresponding: true
-    affiliation: "1, 2, 5"
+    affiliation: "1, 2"
   - name: Milan Klöwer
     orcid: 0000-0002-3920-4356
     affiliation: 3
@@ -22,8 +21,6 @@ affiliations:
    index: 3
  - name: Max Planck Institute for Meteorology, Hamburg, Germany
    index: 4
- - name: Now at Danish Meteorological Institute, Copenhagen, Denmark
-   index: 5
 date: 22 September 2025
 bibliography: paper.bib
 ---
