### 2025-11-25 a942299
use non-breaking space
