### 2025-12-02 d2a9093
Write out countries in affiliation
