### 2025-12-02 3a26385
Minor punctuation fix in paper
