### 2025-12-02 d34accd
Minor fixes for publication list
