+++ paper.bib
@@ -15,7 +15,7 @@
 }
 
 @unpublished{schulz_xbitinfo_2022-1,
-  title = {Xbitinfo: {{Compress}} Datasets Based on Their Information Content},
+  title = {Xbitinfo: Compress Datasets Based on Their Information Content},
   shorttitle = {Xbitinfo},
   author = {Schulz, Hauke and Spring, Aaron and {Klöwer}, Milan},
   date = {2022-10-27},
@@ -29,7 +29,7 @@
   urldate = {2024-12-29},
   abstract = {Information technology — Coding of moving pictures and associated audio for digital storage media at up to about 1,5 Mbit/s — Part 3: Audio},
   langid = {english},
-  author = {International Organization for Standardization},
+  author = {{International Organization for Standardization}},
   year = {1993},
   organization = {ISO}
 }
@@ -41,7 +41,7 @@
   urldate = {2024-12-29},
   abstract = {Information technology — Digital compression and coding of continuous-tone still images: Requirements and guidelines},
   langid = {english},
-  author = {International Organization for Standardization},
+  author = {{International Organization for Standardization}},
   year = {1994},
   organization = {ISO}
 }
@@ -85,7 +85,7 @@
 }
 
 @software{netcdf,
-author = {{NSF Unidata} and Davis, Glenn and Rew, Russ and Heimbigner, Dennis and Hartnett, Edward and Fisher, Ward and Others, Many},
+author = {{NSF Unidata} and Davis, Glenn and Rew, Russ and Heimbigner, Dennis and Hartnett, Edward and Fisher, Ward and others},
 doi = {10.5065/D6H70CW6},
 license = {BSD-3-Clause},
 title = {{NetCDF-C }},
@@ -99,7 +99,7 @@ doi = {10.5334/jors.148},
 journal = {Journal of Open Research Software},
 month = apr,
 number = {1},
-title = {{xarray: N-D labeled Arrays and Datasets in Python}},
+title = {xarray: {N-D} labeled Arrays and Datasets in {Python}},
 volume = {5},
 year = {2017}
 }
