### 2025-12-02 5c7ff0d
Final edits for JOSS paper (#328)
* Write out countries in affiliation

* Minor punctuation fix in paper

* Minor fixes for publication list
