### 2025-12-02 #9178 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
